Time: 0 ms (100.00%), Space: 41.8 MB (74.11%) - LeetHub

+++ 1822-sign-of-the-product-of-an-array/1822-sign-of-the-product-of-an-array.java
@@ -1,6 +1,6 @@
 class Solution {
     public int arraySign(int[] nums) {
-        int temp = 1;
+        byte temp = 1;
         for (int i = 0; i < nums.length; i++) {
             if (nums[i] < 0)
                 temp *= -1;
